more fixes after review

+++ CRUD2-Hibernate/src/main/java/dao/DAOImpl.java
@@ -145,7 +145,7 @@ public class DAOImpl implements DAO {
         Transaction tx1 = null;
         try {
             tx1 = session.beginTransaction();
-            session.update(employee);
+            if (employee != null) session.update(employee);
             tx1.commit();
         } catch (Exception e) {
             if (tx1 != null) tx1.rollback();
@@ -229,19 +229,23 @@ public class DAOImpl implements DAO {
         }
     }
 
-    public void deleteEmployeeById(int id) {
-        Session session = HibernateSessionFactoryUtil.getSessionFactory().openSession();
+    public Employee deleteEmployeeById(int id) {
         Transaction tx1 = null;
-        try {
+        Employee employee = null;
+        try (Session session = HibernateSessionFactoryUtil.getSessionFactory().openSession()) {
             tx1 = session.beginTransaction();
-            Employee employee = session.get(Employee.class, id);
-            session.delete(employee);
+            employee = session.get(Employee.class, id);
+            if (employee != null) {
+                session.delete(employee);
+                System.out.println("Employee " + id + " deleted!");
+            }
             tx1.commit();
         } catch (Exception e) {
             if (tx1 != null) tx1.rollback();
             e.printStackTrace();
-        } finally {
-            session.close();
         }
+        return employee;
     }
 }
+
+
